+++ R/theme_gtsummary_roche.R
@@ -4,10 +4,16 @@
 #' A gtsummary theme for Roche tables
 #' - flextable- and gt-printed tables are styled with reduced padding and font size.
 #' - Uses `label_roche_pvalue()` as the default formatting function for all p-values.
+#' - Uses `label_roche_percent()` as the default formatting function for all percent values.
 #' - Font size defaults are 8 points for all the table by the footers that are 7 points.
 #' - Border defaults to `flextable::fp_border_default(width = 0.5)`.
 #' - The `add_overall(col_label)` default value has been updated.
+#' - The results from `gtsummary::tbl_hierachrical()` and `gtsummary::tbl_hierachrical_count()`
+#'   are now post-processed with `gtsummary::remove_footnote_header()`,
+#'   `crane::modify_zero_recode()`, and `crane::modify_header_rm_md()`.
 #'
+#'
+#' @inheritParams gtsummary::theme_gtsummary_printer
 #' @inheritParams gtsummary::theme_gtsummary_compact
 #'
 #' @return theme list
@@ -24,40 +30,38 @@
 #' )
 #'
 #' reset_gtsummary_theme()
-theme_gtsummary_roche <- function(set_theme = TRUE) {
+theme_gtsummary_roche <- function(font_size = NULL,
+                                  print_engine = c("flextable", "gt", "kable", "kable_extra", "huxtable", "tibble"),
+                                  set_theme = TRUE) {
+  print_engine <- arg_match(print_engine)
+  check_scalar_logical(set_theme)
+
   # {crane} defaults
-  font_size <- 8
-  font_size_gt <- 13 # gt counts size points differently
   border <- flextable::fp_border_default(width = 0.5)
 
   # Initialization with compact gt options -------------------------------------
-  lst_theme <-
-    list(
-      "pkgwide-str:theme_name" = "Roche",
-      # compact gt tables
-      "as_gt-lst:addl_cmds" = list(
-        tab_spanner = rlang::expr(
-          gt::tab_options(
-            table.font.size = !!font_size_gt,
-            data_row.padding = gt::px(1),
-            summary_row.padding = gt::px(1),
-            grand_summary_row.padding = gt::px(1),
-            footnotes.padding = gt::px(1),
-            source_notes.padding = gt::px(1),
-            row_group.padding = gt::px(1)
-          )
-        )
-      )
-    )
+  lst_theme <- list("pkgwide-str:theme_name" = "Roche")
 
   # updating with some pharma-specific bits ------------------------------------
   lst_theme <- lst_theme |>
     append(
       list(
-        "tbl_summary-fn:percent_fun" = gtsummary::label_style_number(digits = 1, scale = 100),
+        "tbl_summary-fn:percent_fun" = label_roche_percent(),
         "pkgwide-fn:pvalue_fun" = label_roche_pvalue(),
         "add_overall.tbl_summary-arg:col_label" = "All Participants  \nN = {gtsummary::style_number(N)}",
-        "pkgwide-str:print_engine" = "flextable"
+        "pkgwide-str:print_engine" = print_engine,
+        "tbl_hierarchical-fn:addnl-fn-to-run" =
+          \(x) {
+            gtsummary::remove_footnote_header(x) |>
+              modify_zero_recode() |>
+              modify_header_rm_md()
+          },
+        "tbl_hierarchical_count-fn:addnl-fn-to-run" =
+          \(x) {
+            gtsummary::remove_footnote_header(x) |>
+              modify_zero_recode() |>
+              modify_header_rm_md()
+          }
       )
     )
 
@@ -68,8 +72,8 @@ theme_gtsummary_roche <- function(set_theme = TRUE) {
       list(
         fontsize =
           list(
-            rlang::expr(flextable::fontsize(size = !!font_size, part = "all")),
-            rlang::expr(flextable::fontsize(size = !!(font_size - 1), part = "footer"))
+            rlang::expr(flextable::fontsize(size = !!(font_size %||% 8), part = "all")),
+            rlang::expr(flextable::fontsize(size = !!((font_size %||% 8) - 1), part = "footer"))
           ),
         border = list(
           rlang::expr(flextable::border_outer(part = "body", border = !!border)),
@@ -96,7 +100,7 @@ theme_gtsummary_roche <- function(set_theme = TRUE) {
             rlang::expr(gt::opt_table_font(font = "arial")),
             rlang::expr(
               gt::tab_options(
-                table.font.size = 13,
+                table.font.size = !!(font_size %||% 13),
                 data_row.padding = gt::px(1),
                 summary_row.padding = gt::px(1),
                 grand_summary_row.padding = gt::px(1),